+++ tests/testthat/test-flatten.R
@@ -5,3 +5,9 @@ test_that("equivalent to paste with collapse", {
 test_that("collapse must be single string", {
   expect_error(str_flatten("A", c("a", "b")), "single string")
 })
+
+test_that("last optionally used instead of final separator", {
+  expect_equal(str_flatten(letters[1:3], ", ", ", and "), "a, b, and c")
+  expect_equal(str_flatten(letters[1:2], ", ", ", and "), "a, and b")
+  expect_equal(str_flatten(letters[1], ", ", ", and "), "a")
+})